:bug: fix integer size

+++ backend/src/main/java/org/volcampanion/entities/RoomEntity.java
@@ -15,7 +15,7 @@ public class RoomEntity {
     private Long id;
     private String name;
     private String description;
-    private Integer capacity;
+    private Short capacity;
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "conference_id")
     private ConferenceEntity conference;
